requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.3
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==2.4.7
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
tzlocal==2.1

# Testing
Telethon==1.21.1
pytest==6.2.4
pytest-asyncio==0.15.1

# code style
black
isort
flake8